+++ CHANGELOG.md
@@ -22,6 +22,7 @@ For a full diff see [`c0c63bb...master`][c0c63bb...master].
 * Renamed `Ergebnis\Environment\Test` to `Ergebnis\Environment\TestVariables` ([#9]), by [@localheinz]
 * Started throwing `Ergebnis\Environment\Exception\CouldNotSet` when a system environment variable could not be set ([#14]), by [@localheinz]
 * Started throwing `Ergebnis\Environment\Exception\CouldNotUnset` when a system environment variable could not be unset ([#15]), by [@localheinz]
+* Started throwing `Ergebnis\Environment\Exception\NotSet` when attempting to retrieve the value of an environment variable that is not set ([#16]), by [@localheinz]
 
 [c0c63bb...master]: https://github.com/ergebnis/environment-variables/compare/c0c63bb...master
 
@@ -34,5 +35,6 @@ For a full diff see [`c0c63bb...master`][c0c63bb...master].
 [#9]: https://github.com/ergebnis/environment-variables/pull/9
 [#14]: https://github.com/ergebnis/environment-variables/pull/14
 [#15]: https://github.com/ergebnis/environment-variables/pull/15
+[#16]: https://github.com/ergebnis/environment-variables/pull/16
 
 [@localheinz]: https://github.com/localheinz